+++ entitled/ibm-security-solutions-prod/RELEASENOTES.md
@@ -1,6 +1,6 @@
 
-# What’s new in Chart Version 1.0
-Initial release of ibm-security-solutions-prod
+# What’s new in Chart Version 1.0.4
+Final version of fix pack 2 for Cloud Pak for Security
 
 
 # Fixes
@@ -18,4 +18,5 @@ Product fixes to support Cloud Pak for Security
 | ----- | ---- | ------------ | ------------------ | ---------------- | ------- |
 | 1.0.0 | Sept 26, 2019|  >=1.11 | | None  | New product release |
 | 1.0.1| Dec 13, 2019|  >=1.11 | | None  |Product fixes to support Cloud Pak for Security|
-| 1.0.3| Feb 5, 2020|  >=1.11 | | None  |Retagging Couchv3 image to pull from local registry|
+| 1.0.4| Feb 11, 2020|  >=1.11 | | None  |Product fixes to support Cloud Pak for Security|
+